Splash + Auth + Redirection + Consent Retrival
The splash screen which leads to the auth screen is now complete.
The auth screen perfroms the redirect which then waits for the user to return to the app.
Once the user is back into the app, data retrival begins.
The user is informed of the data retrival status and then redirected to the auth screen or the home screen as required.4
Relevant assets have been added as well.